+++ pub/patchwork/lib/patchwork.dart
@@ -2,6 +2,9 @@ export 'src/error.dart' show PatchworkException;
 export 'src/model.dart'
     show
         AppliedPatch,
+        CleanupChange,
+        CleanupChangeKind,
+        CleanupResult,
         PackageSource,
         PatchProblem,
         PatchRef,